changes needed for new RoadSigns plugin
git-svn-id: http://josm.openstreetmap.de/svn/trunk@3338 0c6e7542-c601-0410-84e7-c038aed88b3b

+++ src/org/openstreetmap/josm/tools/GBC.java
@@ -101,7 +101,7 @@ public class GBC extends GridBagConstraints {
      * @param weighty   Weight in vertical direction
      * @return This constraint for chaining.
      */
-    public GBC weight(int weightx, int weighty) {
+    public GBC weight(double weightx, double weighty) {
         this.weightx = weightx;
         this.weighty = weighty;
         return this;
